requirements.txt
jmcomic>=2.5.35
Pillow>=10.0.0
PyYAML>=6.0.0